Switched to smart pointers (Fixed memory leaks)
I made the program use smart pointers. The map holds a 2D array of unique pointers to Tile objects. The Tile objects hold smart pointers to ControlNodes. This is because some Tiles will have the same ControlNodes as other Tiles. The ControlNodes hold unique pointers to Nodes.
Also made the rest of the project work with smart pointers where appropriate.
This all fixed the memory leaks I had. (They were happening because of the ControlNodes)